requirements.txt
action-msgs==1.2.1
actionlib-msgs==4.2.4
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
argcomplete==1.8.1
asgiref==3.5.0
attrs==21.2.0
autobahn==24.4.2
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
builtin-interfaces==1.2.1
catkin-pkg-modules==1.0.0
certifi==2024.8.30
channels==4.2.0
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.1
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
daphne==4.1.2
dbus-python==1.2.18
diagnostic-msgs==4.2.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.17.1
domain-coordinator==0.10.0
empy==3.3.4
flake8==4.0.1
Flask==3.1.0
geometry-msgs==4.2.4
gyp==0.1
h11==0.13.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
keyring==23.5.0
lark==1.1.1
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
lxml==4.8.0
MarkupSafe==2.0.1
mccabe==0.6.1
message-filters==4.3.5
more-itertools==8.10.0
nav-msgs==4.2.4
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
osrf-pycommon==2.0.2
packaging==21.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycryptodomex==3.11.0
pycurl==7.44.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
pyxattr==0.7.2
PyYAML==5.4.1
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
requests==2.32.3
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rospkg-modules==1.5.1
rpyutils==0.2.1
SecretStorage==3.3.1
sensor-msgs==4.2.4
service-identity==18.1.0
shape-msgs==4.2.4
six==1.16.0
snowballstemmer==2.2.0
sos==4.7.2
soupsieve==2.3.1
sqlparse==0.5.2
sros2==0.10.5
ssh-import-id==5.11
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
systemd-python==234
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
toml==0.10.2
tomli==2.1.0
trajectory-msgs==4.2.4
Twisted==22.1.0
txaio==23.1.1
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unique-identifier-msgs==2.2.1
urllib3==2.2.3
uvicorn==0.15.0
visualization-msgs==4.2.4
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.0.0
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==5.4.0
